tp2: split les 2 components distinct vue, =
