+++ src/App.vue
@@ -5,7 +5,23 @@
     <router-link to="/add">Add</router-link>
   </nav>
   <router-view/>
+
+  <MyFooter>
+    <template #left>
+      <p>Shahzad Abdul Rahman</p>
+    </template>
+    <template #center>
+      <p>Master TNSID</p>
+    </template>
+    <template #right>
+      <p>Technologie web TP2</p>
+    </template>
+  </MyFooter>
 </template>
+
+<script setup>
+import MyFooter from './components/footer.vue';
+</script>
 
 <style lang="scss">
 #app {
