+++ src/views/AddView.vue
@@ -1,3 +1,9 @@
 <template>
-    
+    <div>
+    <AddBiere />
+  </div>
 </template>
+
+<script setup>
+import AddBiere from '../components/addbiere.vue';
+</script>
