+++ src/views/ListingView.vue
@@ -1,3 +1,10 @@
 <template>
-    
+    <div>
+    <ListingBiere />
+  </div>
 </template>
+
+<script setup>
+import ListingBiere from '../components/listbiere.vue';
+
+</script>
